get data from axios request
