+++ server/index.js
@@ -4,6 +4,8 @@ const overview = require('./Overview/api.js');
 const relatedProducts = require('./RelatedProducts/api.js');
 const qa = require('./QA/api.js');
 const reviews = require('./Reviews/api.js');
+//const api = require('../Reviews/api.js')
+
 
 const app = express();
 const PORT = 3000;
@@ -15,15 +17,12 @@ app.listen(PORT, () => {
   console.log(`Server listening on port: ${PORT}`);
 })
 
-app.get('/products', (req, res) => {
-  console.log('Products', req.body);
-  //var{id} = req.body.id;
-  // api.getProducts(id)
-  //     .then((data) => {
-  //         console.log(data);
-  //         res.status(200).send(data);
-  //     })
-  //     .catch((err) => {
-  //         res.status(404).send(err);
-  //     })
+app.get('/reviews', (req, res) => {
+  reviews.getReviews(req.query.product_id)
+      .then((data) => {
+          res.status(200).send(data.data);
+      })
+      .catch((err) => {
+          res.status(404).send(err);
+      })
 })
